requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==4.0.3
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-environ==0.8.1
django-timezone-field==5.0
gunicorn==20.1.0
idna==3.3
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.30
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2022.4
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.14.1
